+++ src/index.mjs
@@ -1,0 +1,3 @@
+import PublishSubscribe from "./core/PublishSubscribe";
+
+export { PublishSubscribe };