+++ src/utils/clone.mjs
@@ -1,0 +1,9 @@
+import cloneDeep from "./cloneDeep";
+import isPrimitive from "./isPrimitive";
+
+export default function clone(value) {
+  if (isPrimitive(value)) {
+    return value;
+  }
+  return cloneDeep(value);
+}