+++ src/utils/getPrototypeName.mjs
@@ -1,0 +1,3 @@
+export default function getPrototypeName(value) {
+  return Object.prototype.toString.call(value);
+}