+++ src/utils/isObjectEmpty.mjs
@@ -1,0 +1,5 @@
+import objectSize from "./objectSize";
+
+export default function isObjectEmpty(object) {
+  return objectSize(object) === 0;
+}