+++ src/utils/isSame.mjs
@@ -1,0 +1,3 @@
+export default function isSame(variable1, variable2) {
+  return Object.is(variable1, variable2);
+}