+++ src/utils/objectSize.mjs
@@ -1,0 +1,5 @@
+import objectKeys from "./objectKeys";
+
+export default function objectSize(object) {
+  return objectKeys(object).length;
+}